pyproject.toml
[tool.poetry]
name = "google-cloud-gui"
version = "1.0.0"
description = "Google Cloud GUI"
authors = ["Drsanwujiang <drsanwujiang@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/drsanwujiang/google-cloud-gui"
packages = [
    { include = "main.py" }
]
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pyside6 = "^6.6.1"
qt-material = "^2.14"
google-cloud-compute = "^1.15.0"
paramiko = "^3.4.0"
pyte = "^0.8.2"
cryptography = "^41.0.7"

[tool.poetry.group.dev.dependencies]
nuitka = "^1.9.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
